# backend/product/views.py
from __future__ import annotations

from django.db.models import Q
from drf_spectacular.utils import OpenApiParameter, extend_schema
from rest_framework.generics import ListAPIView, RetrieveAPIView
from rest_framework.permissions import AllowAny
from rest_framework.response import Response
from rest_framework.views import APIView

from content.models import CategoryMedicament, CategoryMaladies, CategoryPharmacologie, CategoryTheme, MicroArticlePage
from content.serializers import MicroArticleCardSerializer, TaxonomyResolveResponseSerializer
from learning.models import LessonProgress

from .pagination import FeedCursorPagination
from .serializers import FeedItemSerializer, MicroDetailSerializer


_PRODUCT_CARD_FIELDS = (
    *MicroArticleCardSerializer.default_fields,
    "tags_payload",
    "categories_theme_payload",
    "categories_maladies_payload",
    "categories_medicament_payload",
    "categories_pharmacologie_payload",
)


def _product_card(page: MicroArticlePage, *, include_questions: bool = False) -> dict:
    fields = (*_PRODUCT_CARD_FIELDS, "questions") if include_questions else _PRODUCT_CARD_FIELDS
    data = dict(MicroArticleCardSerializer(page, fields=fields).data)
    data["tags"] = data.pop("tags_payload")
    data["categories_theme"] = data.pop("categories_theme_payload")
    data["categories_maladies"] = data.pop("categories_maladies_payload")
    data["categories_medicament"] = data.pop("categories_medicament_payload")
    data["categories_pharmacologie"] = data.pop("categories_pharmacologie_payload")
    return data


def _taxonomy_model(taxonomy: str):
    if taxonomy == "theme":
        return CategoryTheme, "categories_theme"
    if taxonomy == "maladies":
        return CategoryMaladies, "categories_maladies"
    if taxonomy == "medicament":
        return CategoryMedicament, "categories_medicament"
    if taxonomy == "pharmacologie":
        return CategoryPharmacologie, "categories_pharmacologie"
    return None, None


def _apply_tree_filter(qs, *, taxonomy: str, node_id: int, scope: str):
    model, rel = _taxonomy_model(taxonomy)
    if model is None:
        return qs, False

    node = model.objects.filter(id=node_id).first()
    if node is None:
        return qs, False

    if scope == "exact":
        return qs.filter(**{rel: node}), True

    if scope == "subtree":
        return qs.filter(**{f"{rel}__path__startswith": node.path}), True

    return qs, False


def _parse_int(value: str | None) -> int | None:
    if value is None or value == "":
        return None
    try:
        return int(value)
    except ValueError:
        return None


def _progress_map(user, lesson_ids: list[int]) -> dict[int, dict]:
    if not user.is_authenticated or not lesson_ids:
        return {}
    rows = LessonProgress.objects.filter(user=user, lesson_id__in=lesson_ids)
    return {
        r.lesson_id: {
            "seen": r.seen,
            "completed": r.completed,
            "percent": r.percent,
            "time_ms": r.time_ms,
            "score_best": r.score_best,
            "score_last": r.score_last,
            "updated_at": r.updated_at,
            "last_seen_at": r.last_seen_at,
        }
        for r in rows
    }


class FeedView(ListAPIView):
    permission_classes = [AllowAny]
    pagination_class = FeedCursorPagination
    serializer_class = FeedItemSerializer

    def get_queryset(self):
        qs = (
            MicroArticlePage.objects.live()
            .public()
            .select_related("cover_image")
            .prefetch_related(
                "tags",
                "categories_theme",
                "categories_maladies",
                "categories_medicament",
                "categories_pharmacologie",
            )
            .order_by("-first_published_at", "-id")
        )

        q = self.request.query_params.get("q")
        if q:
            qs = qs.filter(Q(title__icontains=q) | Q(answer_express__icontains=q))

        tags = self.request.query_params.get("tags")
        if tags:
            tag_slugs = [t.strip() for t in tags.split(",") if t.strip()]
            if tag_slugs:
                qs = qs.filter(tags__slug__in=tag_slugs)

        cat_p = self.request.query_params.get("category_theme")
        if cat_p:
            qs = qs.filter(categories_theme__slug=cat_p)

        cat_m = self.request.query_params.get("category_maladies")
        if cat_m:
            qs = qs.filter(categories_maladies__slug=cat_m)

        cat_c = self.request.query_params.get("category_medicament")
        if cat_c:
            qs = qs.filter(categories_medicament__slug=cat_c)

        cat_ph = self.request.query_params.get("category_pharmacologie")
        if cat_ph:
            qs = qs.filter(categories_pharmacologie__slug=cat_ph)

        used_tree_filter = False

        exact_p = _parse_int(self.request.query_params.get("category_theme_exact"))
        if exact_p is not None:
            qs, used_tree_filter = _apply_tree_filter(
                qs, taxonomy="theme", node_id=exact_p, scope="exact"
            )

        subtree_p = _parse_int(self.request.query_params.get("category_theme_subtree"))
        if subtree_p is not None:
            qs, used_tree_filter = _apply_tree_filter(
                qs, taxonomy="theme", node_id=subtree_p, scope="subtree"
            )

        exact_m = _parse_int(self.request.query_params.get("category_maladies_exact"))
        if exact_m is not None:
            qs, used_tree_filter = _apply_tree_filter(
                qs, taxonomy="maladies", node_id=exact_m, scope="exact"
            )

        subtree_m = _parse_int(self.request.query_params.get("category_maladies_subtree"))
        if subtree_m is not None:
            qs, used_tree_filter = _apply_tree_filter(
                qs, taxonomy="maladies", node_id=subtree_m, scope="subtree"
            )

        exact_c = _parse_int(self.request.query_params.get("category_medicament_exact"))
        if exact_c is not None:
            qs, used_tree_filter = _apply_tree_filter(
                qs, taxonomy="medicament", node_id=exact_c, scope="exact"
            )

        subtree_c = _parse_int(self.request.query_params.get("category_medicament_subtree"))
        if subtree_c is not None:
            qs, used_tree_filter = _apply_tree_filter(
                qs, taxonomy="medicament", node_id=subtree_c, scope="subtree"
            )

        exact_ph = _parse_int(self.request.query_params.get("category_pharmacologie_exact"))
        if exact_ph is not None:
            qs, used_tree_filter = _apply_tree_filter(
                qs, taxonomy="pharmacologie", node_id=exact_ph, scope="exact"
            )

        subtree_ph = _parse_int(self.request.query_params.get("category_pharmacologie_subtree"))
        if subtree_ph is not None:
            qs, used_tree_filter = _apply_tree_filter(
                qs, taxonomy="pharmacologie", node_id=subtree_ph, scope="subtree"
            )

        taxonomy = self.request.query_params.get("taxonomy")
        category = _parse_int(self.request.query_params.get("category"))
        scope = self.request.query_params.get("scope")
        if not used_tree_filter and taxonomy and category is not None and scope:
            qs, _ = _apply_tree_filter(qs, taxonomy=taxonomy, node_id=category, scope=scope)

        return qs.distinct()

    def list(self, request, *args, **kwargs):
        page = self.paginate_queryset(self.get_queryset())
        ids = [p.id for p in page]
        progress = _progress_map(request.user, ids)

        data = []
        for item_page in page:
            item = _product_card(item_page)
            item["progress"] = progress.get(item_page.id)
            data.append(item)

        serializer = self.get_serializer(data, many=True)
        return self.get_paginated_response(serializer.data)


class MicroBySlugView(RetrieveAPIView):
    permission_classes = [AllowAny]
    serializer_class = MicroDetailSerializer
    lookup_field = "slug"

    def get_queryset(self):
        return (
            MicroArticlePage.objects.live()
            .public()
            .select_related("cover_image")
            .prefetch_related(
                "tags",
                "categories_theme",
                "categories_maladies",
                "categories_medicament",
                "categories_pharmacologie",
                "microarticle_questions__question",
            )
        )

    def retrieve(self, request, *args, **kwargs):
        page: MicroArticlePage = self.get_object()
        progress = _progress_map(request.user, [page.id]).get(page.id)

        data = {
            **_product_card(page, include_questions=True),
            "links": [b.value for b in page.links] if page.links else [],
            "see_more": [{"type": b.block_type, "value": b.value} for b in page.see_more] if page.see_more else [],
            "progress": progress,
        }
        serializer = self.get_serializer(data)
        return Response(serializer.data)


class MicroByIdView(MicroBySlugView):
    lookup_field = "id"
    lookup_url_kwarg = "id"


class CategoryResolveView(APIView):
    permission_classes = [AllowAny]

    @extend_schema(
        operation_id="legacy_category_resolve",
        parameters=[
            OpenApiParameter(name="taxonomy", type=str, required=True),
            OpenApiParameter(name="path", type=str, required=True),
        ],
        responses=TaxonomyResolveResponseSerializer,
    )
    def get(self, request):
        taxonomy = request.query_params.get("taxonomy")
        path = request.query_params.get("path")
        if not taxonomy or not path:
            return Response({"detail": "taxonomy and path are required."}, status=400)

        model, _ = _taxonomy_model(taxonomy)
        if model is None:
            return Response({"detail": "Unknown taxonomy."}, status=400)

        parts = [p for p in path.split("/") if p]
        if not parts:
            return Response({"detail": "Invalid path."}, status=400)

        node = model.objects.filter(slug=parts[0], depth=1).first()
        if node is None:
            return Response({"detail": "Not found."}, status=404)

        for slug in parts[1:]:
            node = node.get_children().filter(slug=slug).first()
            if node is None:
                return Response({"detail": "Not found."}, status=404)

        breadcrumb_nodes = node.get_ancestors(inclusive=True)
        breadcrumb = [{"id": n.id, "name": n.name, "slug": n.slug} for n in breadcrumb_nodes]
        canonical_path = "/".join([n.slug for n in breadcrumb_nodes])

        return Response(
            {
                "taxonomy": taxonomy,
                "node_id": node.id,
                "breadcrumb": breadcrumb,
                "canonical_path": canonical_path,
            }
        )