# openapi/schema.yaml
openapi: 3.0.3
info:
  title: PharmaPocket API
  version: 1.0.0
  description: Contrat de l'API v1 consommée par le frontend PharmaPocket.
paths:
  /api/v1/auth/account/:
    get:
      operationId: auth_account
      tags:
      - auth
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSummary'
          description: ''
    patch:
      operationId: auth_account_update
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedAccountUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedAccountUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedAccountUpdate'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSummary'
          description: ''
  /api/v1/auth/account/delete/:
    post:
      operationId: auth_account_delete
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAccount'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DeleteAccount'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DeleteAccount'
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/auth/csrf/:
    get:
      operationId: auth_csrf
      tags:
      - auth
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailResponse'
          description: ''
  /api/v1/auth/me/:
    get:
      operationId: auth_me
      tags:
      - auth
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
          description: ''
  /api/v1/auth/preferences/:
    get:
      operationId: auth_preferences
      tags:
      - auth
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
          description: ''
    patch:
      operationId: auth_preferences_update
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedUserPreferencesUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedUserPreferencesUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedUserPreferencesUpdate'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
          description: ''
  /api/v1/categories/resolve/:
    get:
      operationId: legacy_category_resolve
      parameters:
      - in: query
        name: path
        schema:
          type: string
        required: true
      - in: query
        name: taxonomy
        schema:
          type: string
        required: true
      tags:
      - categories
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyResolveResponse'
          description: ''
...
  /api/v1/tags/:
    get:
      operationId: tag_list
      parameters:
      - in: query
        name: limit
        schema:
          type: integer
      - in: query
        name: q
        schema:
          type: string
      tags:
      - tags
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagPayload'
          description: ''
...
components:
  schemas:
    AccountSummary:
      type: object
      properties:
        email:
          oneOf:
          - type: string
            format: email
          - type: string
            maxLength: 0
        username:
          type: string
        pseudo:
          type: string
        has_usable_password:
          type: boolean
      required:
      - email
      - has_usable_password
      - pseudo
      - username
...
    CurrentUser:
      type: object
      properties:
        id:
          type: integer
        email:
          oneOf:
          - type: string
            format: email
          - type: string
            maxLength: 0
        username:
          type: string
        pseudo:
          type: string
        has_usable_password:
          type: boolean
        is_staff:
          type: boolean
        is_superuser:
          type: boolean
        landing_redirect_enabled:
          type: boolean
        landing_redirect_target:
          $ref: '#/components/schemas/LandingRedirectTargetEnum'
      required:
      - email
      - has_usable_password
      - id
      - is_staff
      - is_superuser
      - landing_redirect_enabled
      - landing_redirect_target
      - pseudo
      - username
...
    DeleteAccount:
      type: object
      properties:
        password:
          type: string
          writeOnly: true
    DetailResponse:
      type: object
      properties:
        detail:
          type: string
      required:
      - detail
...
    LandingRedirectTargetEnum:
      enum:
      - start
      - discover
      - cards
      - review
      - quiz
      type: string
      description: |-
        * `start` - start
        * `discover` - discover
        * `cards` - cards
        * `review` - review
        * `quiz` - quiz
...
    PatchedAccountUpdate:
      type: object
      properties:
        pseudo:
          type: string
          nullable: true
...
    PatchedUserPreferencesUpdate:
      type: object
      properties:
        landing_redirect_enabled:
          type: boolean
        landing_redirect_target:
          $ref: '#/components/schemas/LandingRedirectTargetEnum'
...
    TagPayload:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
      required:
      - id
      - name
      - slug
    TaxonomyBreadcrumb:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
      required:
      - id
      - name
      - slug
...
    TaxonomyResolveResponse:
      type: object
      properties:
        taxonomy:
          type: string
        node_id:
          type: integer
        breadcrumb:
          type: array
          items:
            $ref: '#/components/schemas/TaxonomyBreadcrumb'
        canonical_path:
          type: string
      required:
      - breadcrumb
      - canonical_path
      - node_id
      - taxonomy
...
    UserPreferences:
      type: object
      properties:
        landing_redirect_enabled:
          type: boolean
        landing_redirect_target:
          $ref: '#/components/schemas/LandingRedirectTargetEnum'
      required:
      - landing_redirect_enabled
      - landing_redirect_target
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid